src/maiagent/requirements/test.txt
# 測試相關套件
-r base.txt

# 測試框架
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0

# 覆蓋率
coverage==7.3.2

# 測試資料工廠
factory-boy==3.3.0
faker==20.1.0

# HTTP 測試
responses==0.24.1

# 測試報告
pytest-html==4.1.1
pytest-json-report==1.5.0

# 效能測試
locust==2.17.0

# 資料庫測試
pytest-postgresql==5.0.0